refactor: redirect ChatRow / ChatState imports to row-types.ts
15 chat-pane modules + 2 test files now type-import ChatRow and ChatState directly from ./row-types instead of via store.ts's re-export. The re-export from store.ts is removed. Mixed-import sites (use-chat-session, use-bash-mode, Chat, etc.) split their type and runtime imports cleanly. No behavior change.
